requirements.txt
Django==5.2
asgiref==3.8.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
pillow==11.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
bleach==6.2.0
webencodings==0.5.1
